requirements.txt
bidict==0.23.1
blinker==1.8.2
click==8.1.7
Flask==3.0.3
Flask-Cors==5.0.0
Flask-SocketIO==5.4.1
h11==0.14.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
python-dotenv==1.0.1
python-engineio==4.10.1
python-socketio==5.11.4
redis==5.2.0
setuptools==75.3.0
simple-websocket==1.1.0
Werkzeug==3.0.6
wheel==0.44.0
wsproto==1.2.0
